NEC working v1

+++ data/divide_data_sortdate.py
@@ -5,12 +5,12 @@
 # 60%, 20%, 20%, for training, validation and test set
 
 import sys
-from sklearn.cross_validation import StratifiedKFold
+from sklearn.model_selection import StratifiedKFold
 import pandas
 import numpy
 
 data_path = sys.argv[1] + '.csv'
-# data_path = 'appcelerator.csv'
+#data_path = 'appcelerator.csv'
 data = pandas.read_csv(data_path).values
 labels = data[:, 3].astype('int64')
 
@@ -19,10 +19,10 @@ validationSize = 20
 testSize = 20
 
 if trainingSize + validationSize + testSize == 100:
-    numData = len(labels)
-    numTrain = (trainingSize * numData) / 100
-    numValidation = (validationSize * numData) / 100
-    numTest = (testSize * numData) / 100
+    numData = int(len(labels))
+    numTrain = int((trainingSize * numData) / 100)
+    numValidation = int((validationSize * numData) / 100)
+    numTest = int((testSize * numData) / 100)
 
     print ("Total data: %s" % numData)
     print ("Training size: %s, validation size: %s, testing size: %s" % (numTrain, numValidation, numTest))
